+++ R/DEL-model.R
@@ -64,6 +64,7 @@ DEL <- function(x, ..., max_area = NULL) {
 #' @export
 DEL.default <- function(x, ..., max_area = NULL) {
   p <- try(silicate::PATH(x), silent = TRUE)
+  #print("DEL.default")
   if (inherits(p, "try-error")) {
     stop("cannot convert 'x' to a PATH, try 'DEL(silicate::SC(x))' rather than 'DEL(x)'")
   }
@@ -92,7 +93,7 @@ DEL.TRI0 <- function(x, ..., max_area = NULL) {
 DEL.SC <- function(x, ..., max_area = NULL)  {
   .check_area(x$vertex$x_, x$vertex$y_, max_area)
 
-
+#print("DEL.SC")
   ## find if any objects have < 3 edges
   edge_per_object_lt <- x$object["object_"] %>%
     dplyr::inner_join(x$object_link_edge, "object_") %>%
@@ -191,6 +192,8 @@ DEL.SC0 <- function(x, ..., max_area = NULL)  {
 #' @export
 #' @importFrom tibble tibble
 DEL.PATH <- function(x, ..., max_area = NULL) {
+ # print("DEL.PATH")
+ # print(sort(x$object$object_))
   .check_area(x$vertex$x_, x$vertex$y_, max_area)
   dots <- list(...)
   dots[["a"]] <- max_area
@@ -229,7 +232,7 @@ DEL.PATH <- function(x, ..., max_area = NULL) {
 #                        triangle_ = rep(triangle[["triangle_"]], each = 3))
 
   meta <- tibble(proj = get_proj(x), ctime = Sys.time())
-
+#browser()
   structure(list(object = x$object, #object_link_triangle = object_link_triangle,
                  triangle = triangle,
                  vertex = vertex,