+++ R/mesh_plot.R
@@ -60,6 +60,9 @@ mesh_plot.mesh3d <-
     if (!is.null(coords)) {
       warning("argument 'coords' is only used for 'mesh_plot(Raster)', ignoring")
     }
+    if (!is.null(x$material$texture)) {
+      warning("mesh object has a texture path, but cannot be displayed in 2D graphics (try plot3d)")
+    }
     if (!is.null(x$ib)) {
       id <- x$ib
     }
@@ -174,7 +177,7 @@ mesh_plot.default <- function(x,  col = NULL, add = FALSE, zlim = NULL, ...,
     warning("argument 'coords' is only used for 'mesh_plot(Raster)', ignoring")
   }
   if (!is.null(crs)) {
-    xx <- try(reproj::reproj(PATH0(x), target = crs), silent  = TRUE)
+    xx <- try(reproj::reproj(silicate::PATH0(x), target = crs), silent  = TRUE)
     if (inherits(xx, "try-error")) {
       stop("unable to reproject 'x' with 'crs'")
     }